fix: continuation of adoptation of new code format and standards
- make test-utils file to .jsx file instead of .js
- fix jest transform pattern in package.json
- add .eslintignore file to ignore storybook.js because that file
is auto generated
